Updating to use bundler for managing the gem
