--- .autotest
@@ -1,1 +1,0 @@
-require 'autotest/growl'
