+++ .rspec
@@ -1,2 +1,2 @@
 --color
---format nested
+--format documentation
