+++ Rakefile
@@ -1,13 +1,7 @@
-require 'jeweler'
-Jeweler::Tasks.new do |gem|
-  gem.name = "rake-performance"
-  gem.summary = %{Hooks into the rake task and display performance times}
-  gem.description = %{Hooks into the rake task and display performance times}
-  gem.email = "andrew.vos@gmail.com"
-  gem.homepage = "http://github.com/AndrewVos/rake-performance"
-  gem.authors = "Andrew Vos"
-  gem.add_development_dependency "rspec"
-  gem.add_development_dependency "rake"
-end
+require "bundler/gem_tasks"
+require "rspec/core/rake_task"
 
-Jeweler::RubygemsDotOrgTasks.new
+RSpec::Core::RakeTask.new(:spec)
+
+task :default => :spec
+
