--- autotest/discover.rb
@@ -1,1 +1,0 @@
-Autotest.add_discovery { "rspec2" }
