+++ lib/rake_performance/time_helper.rb
