+++ lib/rake_performance/version.rb
@@ -1,0 +1,3 @@
+module RakePerformance
+  VERSION = "0.1.2"
+end
