+++ rake-performance.gemspec
@@ -1,52 +1,32 @@
-# Generated by jeweler
-# DO NOT EDIT THIS FILE DIRECTLY
-# Instead, edit Jeweler::Tasks in Rakefile, and run 'rake gemspec'
-# -*- encoding: utf-8 -*-
+# coding: utf-8
+lib = File.expand_path('../lib', __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require 'rake_performance/version'
 
-Gem::Specification.new do |s|
-  s.name = %q{rake-performance}
-  s.version = "0.1.0"
+Gem::Specification.new do |spec|
+  spec.name          = "rake-performance"
+  spec.version       = RakePerformance::VERSION
+  spec.authors       = ["Andrew Vos"]
+  spec.email         = ["andrew.vos@gmail.com"]
 
-  s.required_rubygems_version = Gem::Requirement.new(">= 0") if s.respond_to? :required_rubygems_version=
-  s.authors = ["Andrew Vos"]
-  s.date = %q{2011-01-30}
-  s.description = %q{Hooks into the rake task and display performance times}
-  s.email = %q{andrew.vos@gmail.com}
-  s.files = [
-    ".autotest",
-    ".rspec",
-    "Rakefile",
-    "VERSION",
-    "autotest/discover.rb",
-    "lib/rake_performance.rb",
-    "lib/time_helper.rb",
-    "rake-performance.gemspec",
-    "spec/rake_performance_spec.rb",
-    "spec/time_helper_spec.rb"
-  ]
-  s.homepage = %q{http://github.com/AndrewVos/rake-performance}
-  s.require_paths = ["lib"]
-  s.rubygems_version = %q{1.3.7}
-  s.summary = %q{Hooks into the rake task and display performance times}
-  s.test_files = [
-    "spec/rake_performance_spec.rb",
-    "spec/time_helper_spec.rb"
-  ]
+  spec.summary       = %q{Hooks into the rake task and display performance times}
+  spec.description   = %q{Hooks into the rake task and display performance times}
+  spec.homepage      = "http://github.com/AndrewVos/rake-performance"
 
-  if s.respond_to? :specification_version then
-    current_version = Gem::Specification::CURRENT_SPECIFICATION_VERSION
-    s.specification_version = 3
+  # Prevent pushing this gem to RubyGems.org by setting 'allowed_push_host', or
+  # delete this section to allow pushing this gem to any host.
+  # if spec.respond_to?(:metadata)
+  #   spec.metadata['allowed_push_host'] = "TODO: Set to 'http://mygemserver.com'"
+  # else
+  #   raise "RubyGems 2.0 or newer is required to protect against public gem pushes."
+  # end
 
-    if Gem::Version.new(Gem::VERSION) >= Gem::Version.new('1.2.0') then
-      s.add_development_dependency(%q<rspec>, [">= 0"])
-      s.add_development_dependency(%q<rake>, [">= 0"])
-    else
-      s.add_dependency(%q<rspec>, [">= 0"])
-      s.add_dependency(%q<rake>, [">= 0"])
-    end
-  else
-    s.add_dependency(%q<rspec>, [">= 0"])
-    s.add_dependency(%q<rake>, [">= 0"])
-  end
+  spec.files         = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
+  # spec.bindir        = "exe"
+  # spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
+  spec.require_paths = ["lib"]
+
+  spec.add_development_dependency "bundler", "~> 1.11"
+  spec.add_development_dependency "rspec", "~> 2.0"
+  spec.add_runtime_dependency "rake", "~> 11.0"
 end
-
