+++ spec/spec_helper.rb
@@ -1,0 +1,3 @@
+$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+require 'rake'
+require 'rake_performance'
